backend/tutorials/spring/security/src/test/java/com/example/security/domain/user/restcontroller/UserRestControllerTest.java
package com.example.security.domain.user.restcontroller;

import com.example.security.domain.user.entity.UserEntity;
import com.example.security.domain.user.service.UserService;
import org.hamcrest.CoreMatchers;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.http.ResponseEntity;
import org.springframework.security.test.context.support.WithMockUser;
import org.springframework.test.web.servlet.MockMvc;

import java.util.Arrays;
import java.util.Optional;

import static org.mockito.Mockito.when;
import static org.springframework.http.MediaType.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;

@WebMvcTest(UserRestController.class)
@DisplayName("User rest controller group test")
public class UserRestControllerTest {
    @Autowired
    private MockMvc mockMvc;

    @MockBean
    private UserService userService;

    @Test
    @WithMockUser(roles = "APP_USER")
    @DisplayName("Get users test")
    void getUsers() throws Exception {
        // When
        when(userService.getUsers(Optional.of(1)))
                .thenReturn(ResponseEntity.ok(Arrays.asList(UserEntity.builder().id(1).name("Mike").build())));

        mockMvc.perform(get("/api/users/1").contentType(APPLICATION_JSON))
                .andExpect(status().isOk())
                .andExpect(jsonPath("$[0].id", CoreMatchers.is(1)))
                .andExpect(jsonPath("$[0].name", CoreMatchers.is("Mike")));
    }
}
